@@ -1,37 +1,52 @@
 ﻿using System;
 using System.Numerics;
-using AdaptiveHud;
+using Dalamud.Interface.Utility;
 using Dalamud.Interface.Windowing;
 using ImGuiNET;
 
-namespace SamplePlugin.Windows;
+namespace AdaptiveHud.Windows;
 
 public class ConfigWindow : Window, IDisposable
 {
     private Configuration Configuration;
 
-    public ConfigWindow(Plugin plugin) : base(
-        "A Wonderful Configuration Window",
-        ImGuiWindowFlags.NoResize | ImGuiWindowFlags.NoCollapse | ImGuiWindowFlags.NoScrollbar |
-        ImGuiWindowFlags.NoScrollWithMouse)
+    // We give this window a constant ID using ###
+    // This allows for labels being dynamic, like "{FPS Counter}fps###XYZ counter window",
+    // and the window ID will always be "###XYZ counter window" for ImGui
+    public ConfigWindow(Plugin plugin) : base("AdaptiveHud Configuration")
     {
-        this.Size = new Vector2(232, 75);
-        this.SizeCondition = ImGuiCond.Always;
+        Flags = ImGuiWindowFlags.NoResize | ImGuiWindowFlags.NoCollapse | ImGuiWindowFlags.NoScrollbar |
+                ImGuiWindowFlags.NoScrollWithMouse;
 
-        this.Configuration = plugin.Configuration;
+        Size = new Vector2(250, 100);
+        SizeCondition = ImGuiCond.Always;
+
+        Configuration = plugin.Configuration;
     }
 
     public void Dispose() { }
 
+    public override void PreDraw()
+    {
+    }
+
     public override void Draw()
     {
-        // can't ref a property, so use a local copy
-        var configValue = this.Configuration.SomePropertyToBeSavedAndWithADefault;
-        if (ImGui.Checkbox("Random Config Bool", ref configValue))
+        string[] layoutOptions = ["1", "2", "3", "4"];
+
+        ImGui.SetNextItemWidth(50 * ImGuiHelpers.GlobalScale);
+        int windowedValue = Configuration.LayoutForWindowedMode;
+        if (ImGui.Combo("Windowed mode layout", ref windowedValue, layoutOptions, 4))
         {
-            this.Configuration.SomePropertyToBeSavedAndWithADefault = configValue;
-            // can save immediately on change, if you don't want to provide a "Save and Close" button
-            this.Configuration.Save();
+            Configuration.LayoutForWindowedMode = windowedValue;
+            Configuration.Save();
+        }
+        ImGui.SetNextItemWidth(50 * ImGuiHelpers.GlobalScale);
+        int fullscreenValue = Configuration.LayoutForFullscreenMode;
+        if (ImGui.Combo("Fullscreen mode layout", ref fullscreenValue, layoutOptions, 4))
+        {
+            Configuration.LayoutForFullscreenMode = fullscreenValue;
+            Configuration.Save();
         }
     }
 }
